// nse/src/app/models/nominee-data.model.ts
export interface NomineeData{
    customerID: number;
    investorName: string;
    executeThroughPoa:string;
    bankName: string;
    accountNumber:number;
    branchName: string;
    accountType: string; 
    micrNumber:number;
    ifscCode: string;
    bankHolderName: string;
    bankHolderName1: string;
    bankHolderName2: string;
    achAmount: number;
    achFromDate: Date;
    achToDate: Date;
    modeOfHolder:string;
}